// src/types/myPageInfo.ts
export interface MyPageInfo {
    userId: number;
    userName: string;
    employeeId: string;
    departmentName: string;
    gradeName: string;
    phone: string;
    email: string;
    address: string;
    bank: string;
    accountNum: string;
    startDate: string;
    birth: string;
    profileUrl: string;
    residentNumber: string;
    age: number;
}